Bug fixes, control of visibility of tokens in FU where player has no permissions
If you tried to add a stunt, aspect, track, or skill to a character and it contained a period, it would not be added correctly. If you try this now, the system will replace the period with a blank space and the thing being edited will save properly.

Players can no longer see information in Fate Utilities for any token for which they have permissions None. They will be able to see the aspects and other data for a token thye have Limited or Observer permissions for. Players can always see other players' information.