Drag and drop extra functions added
You can now drag and drop an extra from the character on to a Pick Up Stix chest or another character sheet, compendium, or item list. This will remove it from the character sheet if dragged to pick up stix or another character sheet. but will NOT delete the item if dragged to a compendium or the item list.