Temporarily at least set the Tracks editor to use untoggled prose-mirror instances
Due to the issue I was ahving with the rich text presented by prose-mirror not updating, I have temporarily at least set these to be non-toggled PMs.

For similar reasons I have changed the stunt editor to submit on change and removed the save button.

I think the solution will be to capture the current values of all data to an object that can be maintained in memory when the object is refreshed.